+++ ExpandWorldMusic/ExpandWorldMusic.cs
@@ -12,7 +12,7 @@ public class EWM : BaseUnityPlugin
 {
   public const string GUID = "expand_world_music";
   public const string NAME = "Expand World Music";
-  public const string VERSION = "1.12";
+  public const string VERSION = "1.13";
 
 #nullable disable
   public static CustomSyncedValue<List<Data>> valueMusicData;