+++ ExpandWorldMusic/LocationManager.cs
@@ -11,7 +11,7 @@ public class LocationManager
   public static string ReferenceFileName = "ref_expand_location_music.yaml";
   public static string FileName = "expand_location_music.yaml";
   public static string Pattern = "expand_location_music*.yaml";
-  public static Dictionary<string, LocationData> Data = [];
+  public static Dictionary<int, LocationData> Data = [];
 
   public static bool IsServer() => ZNet.instance && ZNet.instance.IsServer();
   public static void ToFile()
@@ -53,13 +53,14 @@ public class LocationManager
     Data.Clear();
     foreach (var d in data)
     {
-      if (Data.ContainsKey(d.name))
+      var hash = d.name.GetStableHashCode();
+      if (Data.ContainsKey(hash))
       {
         if (IsServer())
           Log.Warning($"Duplicate location music entry for '{d.name}' found.");
         continue;
       }
-      Data[d.name] = d;
+      Data[hash] = d;
     }
   }
   public static void SetupWatcher()
@@ -80,20 +81,57 @@ public class InitializeLocationContent
 }
 
 
-[HarmonyPatch(typeof(MusicLocation), nameof(MusicLocation.Awake))]
-public class MusicLocatioAwake
+[HarmonyPatch]
+public class LocationPatch
 {
-  static void Prefix(MusicLocation __instance)
+  static readonly int ReferenceHash = "locationreference".GetStableHashCode();
+
+  [HarmonyPatch(typeof(LocationProxy), nameof(LocationProxy.SpawnLocation)), HarmonyPostfix]
+  static void SpawnLocation(LocationProxy __instance)
   {
-    var location = Location.GetLocation(__instance.transform.position);
-    var prefab = Utils.GetPrefabName(location?.name ?? __instance.gameObject.name);
-    if (!LocationManager.Data.TryGetValue(prefab, out var data)) return;
-    __instance.m_forceFade = data.forceFade;
-    __instance.m_notIfEnemies = data.notIfEnemies;
-    __instance.m_oneTime = data.oneTime;
-    __instance.m_radius = data.radius;
-    __instance.m_addRadiusFromLocation = data.radiusFromLocation;
-    var audioSource = __instance.GetComponent<AudioSource>();
+    if (!__instance.m_nview) return;
+    // This function will be recalled later if spawn was delayed, so skip for now.
+    if (__instance.m_locationNeedsSpawn) return;
+    var zdo = __instance.m_nview.GetZDO();
+    // Expand World Data adds this so that clones and blueprints can be referenced.
+    var hash = zdo.GetInt(ReferenceHash);
+    if (hash == 0)
+      hash = zdo.GetInt(ZDOVars.s_location);
+    if (!LocationManager.Data.TryGetValue(hash, out var data)) return;
+    var music = EnsureMusic(__instance);
+    if (!music) return;
+    Apply(music, data);
+  }
+
+  private static MusicLocation EnsureMusic(LocationProxy proxy)
+  {
+    var go = proxy.gameObject;
+    // Instance might already have MusicLocation component, so prefer using that.
+    if (proxy.m_instance)
+      go = proxy.m_instance.gameObject;
+    var music = go.GetComponent<MusicLocation>();
+    if (!music)
+    {
+      go.AddComponent<AudioSource>();
+      music = go.AddComponent<MusicLocation>();
+    }
+    if (!music.m_nview)
+    {
+      // Location probably doesn't have ZNetView but LocationProxy has, so bridge that to extend support.
+      music.m_nview = proxy.m_nview;
+      if (music.m_nview)
+        music.m_nview.Register("SetPlayed", music.SetPlayed);
+    }
+    return music;
+  }
+  private static void Apply(MusicLocation music, LocationData data)
+  {
+    music.m_forceFade = data.forceFade;
+    music.m_notIfEnemies = data.notIfEnemies;
+    music.m_oneTime = data.oneTime;
+    music.m_radius = data.radius;
+    music.m_addRadiusFromLocation = data.radiusFromLocation;
+    var audioSource = music.m_audioSource;
     if (!audioSource) return;
     audioSource.volume = data.volume;
     audioSource.loop = data.loop;