+++ publish/CHANGELOG.md
@@ -1,3 +1,7 @@
+- 1.13
+  - Adds support for location clones and Blueprint locations from Expand World data mod.
+  - Fixes music not working for locations that normally don't have music.
+
 - 1.12
   - Adds support for modifying sound effects.
   - Adds support for adding new sound effects.
@@ -14,6 +18,3 @@
 - 1.9
   - Fixes the same location having multiple location entries causing multiple location music entries. Now only the first entry is used.
   - Fixes multiple same location music entries causing an error. Now only the first entry is used.
-
-- 1.8
-  - Adds support for location music.